% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMMIplot.R
\name{AMMIplot}
\alias{AMMIplot}
\title{AMMIplot function}
\usage{
AMMIplot(bilinearObject, plots = "linear", color = c("darkgreen",
  "darkblue"), PC = 1, f = 0.5, pdf.filename = NULL, ...)
}
\arguments{
\item{bilinearObject}{object from output of a call to biliner()}

\item{plots}{character vector of maximum length 2 containing the names of the plots to be produced. Possible arguments are "linear" and "winner". If both are specified in a single call, both plots will be plotted to the same device}

\item{color}{character vector of length 2 containing colors for the plots. The first specifies the genotype color and the second the environment color. If only one color is specified, only genotypes will be colored}

\item{PC}{integer. Principal component to plot. The default is 1}

\item{f}{numeric. Scale parameter in (0, 1) for exponent on eigenvalues for weighting the genotype scores. Environment scores are weighted 1 - f. Default is 0.5}

\item{pdf.filename}{name of pdf file to output.}
}
\description{
This function will make linear and winner plots from a bilinear() object like those in Gauch & Zobel (1997).
}
\details{
Many arguments can be passed on to plot() through (...)
}
\examples{

data(soy)
AMMIfit <- bilinear(x = soyShort)
AMMIplot(AMMIfit)
AMMIplot(AMMIfit, "winner")
AMMIplot(AMMIfit, "winner", color = "hotpink")
AMMIplot(AMMIfit, c("linear", "winner"), color = c("hotpink", "darkorchid"))

}
\keyword{AMMI}
