% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdecomp.R
\name{bdecomp}
\alias{bdecomp}
\title{Expectation maximization algorithm to impute missing GxE cells}
\usage{
bdecomp(Y, model)
}
\arguments{
\item{Y}{matrix containing numeric values of cell means with genotypes on rows and environments on columns}

\item{model}{bilinear model to be fit. Arguments can be "AMMI", "GGE", "SREG", "EGE", "GREG". "GGE" and "SREG" are equivalent, as are "EGE" and "GREG".}
}
\value{
Returns the bilinear decomposition of the GE matrix
}
\description{
This function will impute missing cells in the GxE table using an expectation maximization algorithm. Not intended for direct use by the user.
}
\details{
A: Matrix of additive effects (mu + G + E)
Emat: Matrix of residuals from the additive mode
Edecomp: singlualr value decomposition of Emat
Lambda: singular values from 
nu: degrees of freedom
}
\examples{
data(soy)
bdecomp(soyMeansMat, "AMMI")
}
