% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorateBBplot.R
\name{decorateBBplot}
\alias{decorateBBplot}
\title{decorateBBplot Function}
\usage{
decorateBBplot(Gscores, Escores, Eeffect, axes.names, decorate = TRUE,
  scaledPlot = TRUE, Gnames)
}
\arguments{
\item{Gscores}{matrix of genotype scores}

\item{Escores}{matrix of environment scores}

\item{Eeffect}{numeric vector environment effects(?)}

\item{axes.names}{character vector of axis names}

\item{decorate}{logical, indicates if the environment winenr panels should be colored}

\item{scaledPlot}{logical, indicates if genotypes scores and escores shoudl be scaled to one another (how does f effect this?)}

\item{Gnames}{names of genotypes to be printed, default is outer edge}
}
\value{
list sectors
}
\description{
This function is called by BBplot for drawing GGE biplot with environment winner designations. This function is not intended to be used by the user directly
}
\details{
Uses chull() to compute convex hull of biplot points, draws lines to each of them sucessively, and then draws lines from the origin perpendicular to each segment.
Colors are added to sectors containing genotypes if decorate = TRUE
}
\examples{

data(ontario)
GGEfit <- bilinear(x = ontario, model = "GGE")
BBplot(GGEfit, decorateGGE = TRUE)

}
\keyword{GGE}
