% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheta.R
\name{getTheta}
\alias{getTheta}
\title{Extract Theta from decomposition of GxE residual matrix from additive, for some k in {0, 1, ..., M}.}
\usage{
getTheta(k, Edecomp)
}
\arguments{
\item{k}{number of PCs, k in {0, 1, ..., M}.}

\item{Edecomp}{decomposition of GxE residual matrix from additive}
}
\value{
Returns the matrix Theta of dimension k
}
\description{
Calculate dimension reduction of kth degree
}
\details{
Theta: matrix of dimension k, from decomposition of GxE residual
}
\examples{
data(soyMeanMat)
decomp <- bdecomp(soyMeanMat, "AMMI")
getTheta(k = 2, decomp$Edecomp)

}
