% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontario.R
\docType{data}
\name{ontario}
\alias{ontario}
\title{Onterio wheat data containing 18 genotypes evaluated in 9 environments from Yan et al. (2000)}
\format{#' An object of class \code{"matrix"} containing cell means of genotypes on rows and environments on columns}
\usage{
data(ontario)
}
\description{
Onterio wheat data containing 18 genotypes evaluated in 9 environments from Yan et al. (2000)
}
\examples{
data(ontario)
print(ontario)
}
\references{
Yan, W., Hunt, L. A., Sheng, Q., & Szlavnics, Z. (2000).
Cultivar evaluation and mega-environment investigation based on the GGE biplot. Crop Science, 40(3), 597-605. 
(\href{https://dl.sciencesocieties.org/publications/cs/abstracts/40/3/597}{Crop Science})

Yan, W., Kang, M.S., Ma, B., Woods, S. and Cornelius, P.L. (2007)
GGE biplot vs. AMMI analysis of genotype-by-environment data. Crop Science, 47(2), pp.643-653.
(\href{https://dl.sciencesocieties.org/publications/cs/abstracts/47/2/643}{Crop Science})
}
\keyword{datasets}
