% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soyMeanDf.R
\docType{data}
\name{soyMeanDf}
\alias{soyMeanDf}
\title{Soybean means data from Zobel, Wright and Gauch (1988)}
\format{soyMeanDf An object of class \code{"data.frame"} containing cell means (\code{"y"}) of genotype (\code{"G"}) within environment (\code{"E"})}
\usage{
data(soyMeanDf)
}
\description{
Soybean means data from Zobel, Wright and Gauch (1988)
}
\examples{
data(soyMeanDf)
head(soyMeanDf)
}
\references{
Zobel, R. W., Wright, M. J., & Gauch, H. G. (1988). 
Statistical analysis of a yield trial. Agronomy Journal, 80(3), 388-393.
(\href{https://dl.sciencesocieties.org/publications/aj/abstracts/80/3/AJ0800030388}{Agronomy Journal})
}
\keyword{soy}
