% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soyMeanMat.R
\docType{data}
\name{soyMeanMat}
\alias{soyMeanMat}
\title{Soybean means data from Zobel, Wright and Gauch (1988)}
\format{soyMeanMat An object of class \code{"matrix"} containing cell means of genotypes on rows and environments on columns}
\usage{
data(soyMeanMat)
}
\description{
Soybean means data from Zobel, Wright and Gauch (1988)
}
\examples{
data(soyMeanMat)
print(soyMeanMat)
}
\references{
Zobel, R. W., Wright, M. J., & Gauch, H. G. (1988). 
Statistical analysis of a yield trial. Agronomy Journal, 80(3), 388-393.
(\href{https://dl.sciencesocieties.org/publications/aj/abstracts/80/3/AJ0800030388}{Agronomy Journal})
}
\keyword{soyMeanMat}
