% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormat.R
\name{dataFormat}
\alias{dataFormat}
\title{dataFormat Function}
\usage{
dataFormat(...)
}
\arguments{
\item{...}{arguments to dataFormat()}
}
\value{
formatted data frame for use inside bilinear()
}
\description{
This function is made to format data within bilinear(). Not intended for use by user.
}
\keyword{bilinear}
